Move the with_ridiculous_timeout decorator to the utils module.
